+++ app/src/main/java/com/example/pasteleriaapp/model/Usuario.kt
@@ -10,10 +10,10 @@ import androidx.room.PrimaryKey
 )
 data class Usuario(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val nombre: String,
     val correo: String,
-    val contrasena: String, // En una app real, esto debería estar encriptado
+    val contrasena: String,
     val direccion: String,
     val profilePictureUri: String? = null
 )
